// frontend/src/generated/models/Sequence.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sentinel Messaging Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SequenceStepInput } from './SequenceStepInput';
import {
    SequenceStepInputFromJSON,
    SequenceStepInputFromJSONTyped,
    SequenceStepInputToJSON,
    SequenceStepInputToJSONTyped,
} from './SequenceStepInput';

/**
 * 
 * @export
 * @interface Sequence
 */
export interface Sequence {
    /**
     * 
     */
    id: string;
    /**
     * 
     */
    name: string;
    /**
     * 
     */
    steps: Array<SequenceStepInput>;
}

/**
 * Check if a given object implements the Sequence interface.
 */
export function instanceOfSequence(value: object): value is Sequence {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('steps' in value) || value['steps'] === undefined) return false;
    return true;
}

export function SequenceFromJSON(json: any): Sequence {
    return SequenceFromJSONTyped(json, false);
}

export function SequenceFromJSONTyped(json: any, ignoreDiscriminator: boolean): Sequence {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'steps': ((json['steps'] as Array<any>).map(SequenceStepInputFromJSON)),
    };
}

export function SequenceToJSON(json: any): Sequence {
    return SequenceToJSONTyped(json, false);
}

export function SequenceToJSONTyped(value?: Sequence | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'steps': ((value['steps'] as Array<any>).map(SequenceStepInputToJSON)),
    };
}